handles: reserve some top-100 handles; prevent new 2-char handle registrations (#531)
* handle: disallow new 2-char handle registration

This should not impact existing handles, only new registrations.

We discussed blocking these to prevent possible future URL path
collisions with, eg, ISO country or language codes.

* handle: reserve some more famous handles

This isn't the full list of thousands of twitter handles that have more
than 5 million or more than 1 million followers. Instead, it is only the
top 100 or so, plus a handful of additional top-50 instagram accounts (when
different), and a couple recent famous figures. Pretty subjective.

* handle: overall and segment lengths; enforce ASCII; test dashes

The 'address' package we are using to check domains doesn't seem to do
any overall length checks. The values and examples i'm testing here are
based on general domain limits. Technically a web domain could be 255,
it is email domains limited to 253 chars, but that seems like a
reasonable value to cut off at.