Include description in feed search
This includes the description when searching for popular feeds instead
of just the display name in the getPopularFeedGenerators endpoint.

Our main motivation behind this change is a bit selfish, where right now
searching for “furry” in the popular feed yields no list from
@furryli.st despite having the most popular furry feeds across Bluesky.

Our feeds are named as short as possible to be recognizable (with the
paws emoji 🐾) but to take up as little space as possible in the user’s
feed list.

Some other feeds have similar discoverability problems due to having
branded or short names, such as “My Bangers” (by Jaz), “📽️” (by
Frances), or “GreenSky” (by Ketan) where the title is obvious but
would be easier to discover if the description was included in search.